refactor: use PrimaryButton and LazyColumn in CreateDukan screen
