+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/CreateDukan/content/DukanImageCropContent.kt
@@ -2,10 +2,10 @@ package net.thechance.mena.dukan.presentation.screen.createDukan.content
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -15,6 +15,7 @@ import net.thechance.mena.dukan.presentation.screen.CreateDukan.components.Image
 import net.thechance.mena.dukan.presentation.screen.CreateDukan.components.SaveButton
 import net.thechance.mena.dukan.presentation.screen.CreateDukan.components.UploadAnotherImageButton
 import net.thechance.mena.dukan.presentation.screen.CreateDukan.components.ZoomControls
+import net.thechance.mena.dukan.presentation.util.stubPreviews.PreviewCreateDukanInteractionListener
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanInteractionListener
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanUiState
 import org.jetbrains.compose.ui.tooling.preview.Preview
@@ -24,7 +25,7 @@ fun DukanImageCropContent(
     state: CreateDukanUiState,
     interactionListener: CreateDukanInteractionListener
 ) {
-    Column(
+    LazyColumn(
         modifier = Modifier
             .fillMaxSize()
             .background(Theme.colorScheme.background.surfaceHigh)
@@ -33,58 +34,46 @@ fun DukanImageCropContent(
         verticalArrangement = Arrangement.Top,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        ImageCroppingBox(
-            content = {},
-            modifier = Modifier.padding(top = Theme.spacing._24)
-        )
+        item {
+            ImageCroppingBox(
+                content = {},
+                modifier = Modifier.padding(top = Theme.spacing._24)
+            )
+        }
 
-        ZoomControls(
-            onZoomInClicked = interactionListener::onZoomInClicked,
-            onZoomOutClicked = interactionListener::onZoomOutClicked,
-            onResetClicked = interactionListener::onResetClicked,
-            modifier = Modifier.padding(top = Theme.spacing._12),
-            isZoomOutEnabled = state.isZoomOutEnabled
-        )
+        item {
+            ZoomControls(
+                onZoomInClicked = interactionListener::onZoomInClicked,
+                onZoomOutClicked = interactionListener::onZoomOutClicked,
+                onResetClicked = interactionListener::onResetClicked,
+                modifier = Modifier.padding(top = Theme.spacing._12),
+                isZoomOutEnabled = state.isZoomOutEnabled
+            )
+        }
 
-        SaveButton(
-            onClick = interactionListener::onSaveClicked,
-            modifier = Modifier.padding(top = Theme.spacing._12)
-        )
+        item {
+            SaveButton(
+                onClick = interactionListener::onSaveClicked,
+                modifier = Modifier.padding(top = Theme.spacing._12)
+            )
+        }
 
-        UploadAnotherImageButton(
-            onClick = interactionListener::onUploadAnotherImageClicked,
-            modifier = Modifier.padding(top = Theme.spacing._12)
-        )
+        item {
+            UploadAnotherImageButton(
+                onClick = interactionListener::onUploadAnotherImageClicked,
+                modifier = Modifier.padding(top = Theme.spacing._12)
+            )
+        }
     }
 }
 
 @Preview
 @Composable
-fun DukanImageCropContentPreview() {
+private fun DukanImageCropContentPreview() {
     MenaTheme {
         DukanImageCropContent(
             state = CreateDukanUiState(),
-            interactionListener = object : CreateDukanInteractionListener {
-                override fun onButtonClicked() {}
-
-                override fun onBackClicked() {}
-
-                override fun onClickUploadImage() {}
-
-                override fun onClickEditImage() {}
-
-                override fun onCLickNext() {}
-
-                override fun onSaveClicked() {}
-
-                override fun onZoomInClicked() {}
-
-                override fun onZoomOutClicked() {}
-
-                override fun onResetClicked() {}
-
-                override fun onUploadAnotherImageClicked() {}
-            }
+            interactionListener = PreviewCreateDukanInteractionListener
         )
     }
 }
