+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/util/stubPreviews/PreviewCreateDukanInteractionListener.kt
@@ -1,0 +1,16 @@
+package net.thechance.mena.dukan.presentation.util.stubPreviews
+
+import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanInteractionListener
+
+object PreviewCreateDukanInteractionListener : CreateDukanInteractionListener {
+    override fun onButtonClicked() {}
+    override fun onBackClicked() {}
+    override fun onClickUploadImage() {}
+    override fun onClickEditImage() {}
+    override fun onCLickNext() {}
+    override fun onSaveClicked() {}
+    override fun onZoomInClicked() {}
+    override fun onZoomOutClicked() {}
+    override fun onResetClicked() {}
+    override fun onUploadAnotherImageClicked() {}
+}
